requirements.txt
google-genai
gradio
numpy
opencv-python
torch
transformers
Pillow
ultralytics